requirements.txt
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
filelock==3.17.0
identify==2.6.6
idna==3.10
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
virtualenv==20.29.1
